requirements-dev.txt
-r requirements.txt
pytest-cov==5.0.0
mypy==1.13.0
